requirements.txt
django
django-markdownx
django-reversion
simplejson
